Fix copy tests
